+++ src/collective/dms/mailcontent/tests/test_indexers.py
@@ -35,8 +35,14 @@ class TestIndexers(unittest.TestCase):
                                                **{'title': 'Test 1',
                                                   'sender': RelationValue(self.intids.getId(self.general)),
                                                   'recipients': [RelationValue(self.intids.getId(self.general))]})
+        self.imail1 = createContentInContainer(self.portal, 'dmsincomingmail',
+                                               **{'title': 'Test 1',
+                                                  'sender': [RelationValue(self.intids.getId(self.general)),
+                                                             RelationValue(self.intids.getId(self.divisionalpha))],
+                                                  'recipients': [RelationValue(self.intids.getId(self.general))]})
 
     def test_sender_index(self):
+        # test outgoingmail
         indexer = sender_index(self.omail1)
         # test with held_position
         index_value = indexer()
@@ -47,6 +53,18 @@ class TestIndexers(unittest.TestCase):
         self.omail1.sender = RelationValue(self.intids.getId(self.divisionalpha))
         index_value = indexer()
         self.assertEqual(len(index_value), 4)
+        self.assertIn(self.divisionalpha.UID(), index_value)
+        self.assertIn('l:%s' % self.divisionalpha.UID(), index_value)
+        self.assertIn('l:%s' % self.corpsa.UID(), index_value)
+        self.assertIn('l:%s' % self.armeedeterre.UID(), index_value)
+        # test incomingmail
+        indexer = sender_index(self.imail1)
+        # test with held_position
+        index_value = indexer()
+        self.assertEqual(len(index_value), 5)
+        self.assertIn(self.general.UID(), index_value)
+        self.assertIn('l:%s' % self.armeedeterre.UID(), index_value)
+        # test with organization
         self.assertIn(self.divisionalpha.UID(), index_value)
         self.assertIn('l:%s' % self.divisionalpha.UID(), index_value)
         self.assertIn('l:%s' % self.corpsa.UID(), index_value)